added auth on most routes except three

+++ client/store/storeComponents/updateUser.js
@@ -12,38 +12,40 @@ const _updateUser = (user) => ({
   user,
 });
 
-// thunk
 const updateUser = ({
-  id,
-  firstName,
-  lastName,
-  email,
-  password,
-  address,
+  id, firstName, lastName, email, password, address,
 }) => async (dispatch) => {
   try {
-    const user = (
-      await axios.put(`/api/users/${id}`, {
-        firstName,
-        lastName,
-        email,
-        password,
-        address,
-        headers: getToken().headers,
-      })
-    ).data;
+    const token = window.localStorage.getItem('token');
+    if (!token) {
+      const error = new Error('Unauthorized');
+      throw error;
+    } else {
+      const user = (
+        await axios.put(`/api/users/${id}`, {
+          firstName,
+          lastName,
+          email,
+          password,
+          address,
+          headers: {
+            authorization: token,
+          },
+        })
+      ).data;
 
-    const newToken = (
-      await axios.post('/api/login/auth', {
-        email: user.email,
-        password,
-      })
-    ).data.token;
-    window.localStorage.clear();
-    window.localStorage.setItem('token', newToken);
-    delete user.password;
-    console.log(user);
-    dispatch(_updateUser(user));
+      const newToken = (
+        await axios.post('/api/login/auth', {
+          email: user.email,
+          password,
+        })
+      ).data.token;
+      window.localStorage.clear();
+      window.localStorage.setItem('token', newToken);
+      delete user.password;
+      console.log(user);
+      dispatch(_updateUser(user));
+    }
   } catch (err) {
     console.log(err.response);
   }
